% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchalohhwatersurvey.R
\docType{data}
\name{nchalohhwatersurvey}
\alias{nchalohhwatersurvey}
\title{Nchalo Water Project – Household Survey 2021}
\format{
A tibble with 74 rows and 39 variables
\describe{
\item{submitted_on}{Date and time the survey was submitted}
\item{community_name}{Name of the community surveyed}
\item{latitude}{Latitude coordinate of the survey location}
\item{longitude}{Longitude coordinate of the survey location}
\item{community_population}{Estimated total number of people in the community}
\item{main_livelihood_source}{Primary source of livelihood for the household}
\item{main_livelihood_source_comments}{Additional comments on the households main livelihood}
\item{income_in_2020_mwk}{Estimated household income in 2020 from the main livelihood source (in MWK)}
\item{income_comments}{Additional comments about the reported income}
\item{income_spending_priority}{Main priorities for spending the household income}
\item{main_drinking_water_source}{Primary source of drinking water for the community}
\item{main_drinking_water_source_other}{Specification if Other water source was selected}
\item{uses_borehole_for_other_purposes}{Whether the borehole is used for purposes other than drinking}
\item{other_borehole_use_comments}{Comments on other uses of borehole water}
\item{reasons_not_pursuing_multiple_uses}{Reasons for not using borehole water for multiple purposes}
\item{reasons_not_pursuing_multiple_uses_other}{Specification if an Other reason was selected}
\item{value_addition_opportunity}{Whether value addition opportunities exist for a new water project}
\item{value_addition_opportunity_details}{Details about the identified value addition opportunities}
\item{value_addition_opportunity_other}{Specification if an Other opportunity was mentioned}
\item{land_allocation_responsible_person}{Person or group responsible for land allocation for new projects}
\item{leadership_allocation_decider}{Entity that decides leadership roles in community projects}
\item{leadership_allocation_other}{Specification if an Other decision maker was indicated}
\item{gender_guidelines_for_leadership}{Existence of gender guidelines in community leadership}
\item{gender_guidelines_comments}{Comments on the gender guidelines}
\item{pay_for_water}{Whether the community pays for water usage}
\item{water_payment_amount}{Amount paid for water usage (in MWK)}
\item{water_payment_frequency}{Frequency of water payment}
\item{water_payment_frequency_other}{Specification if an Other frequency was selected}
\item{reason_for_not_paying}{Main reason for not paying for water}
\item{reason_for_not_paying_other}{Specification if an Other reason was given}
\item{nearest_school_type}{Type of school closest to the community}
\item{govt_school_accessibility}{Whether government schools are accessible to all}
\item{nearest_health_facility_type}{Type of health facility closest to the community}
\item{health_facility_accessibility}{Whether the health facility is affordable and accessible to all}
\item{health_facility_accessibility_other}{Specification if an Other accessibility concern was mentioned}
\item{community_solved_social_problem}{Whether the community has independently addressed social problems}
\item{community_solved_social_problem_comments}{Comments on how social problems were addressed by the community}
\item{flow_rate_test_possible_today}{Whether it was possible to conduct a flow rate test during the visit}
\item{has_temp_ec_tds_meter}{Whether the team had a temperature and EC or TDS meter available}
}
}
\usage{
nchalohhwatersurvey
}
\description{
This is a community-level household survey conducted as part of the Nchalo Water Project. Its purpose is to
assess livelihoods, water usage, community governance, infrastructure access, and opportunities for value addition in
rural water supply and WASH-related development intervention.
}
\keyword{datasets}
